\docType{package}
\name{grnn-package}
\alias{grnn}
\alias{grnn-package}
\title{GRNN}
\description{
  General regression neural network.
}
\details{
  The program GRNN implements the algorithm proposed by
  Specht (1991).
}
\author{
  Pierre-Olivier Chasset
}
\references{
  Specht D.F. (1991). A general regression neural network.
  IEEE Transactions on Neural Networks, 2(6):568-576.
}
\keyword{Neural}
\keyword{Regression}
\keyword{network,}

